+ mapper from trips rersponse to daily forecast data
